+++ helpers/WAS_Node_Suite.py
@@ -1,13 +1,4 @@
-CLIPSEG_MODELS = [
-    "models--CIDAS--clipseg-rd64-refined",
-]
-
-
 class WAS_Node_Suite:
-    @staticmethod
-    def models():
-        return CLIPSEG_MODELS
-
     @staticmethod
     def add_weights(weights_to_download, node):
         node_class = node.get("class_type")
@@ -16,17 +7,7 @@ class WAS_Node_Suite:
             node_class == "CLIPSeg Model Loader"
             and model_name == "CIDAS/clipseg-rd64-refined"
         ):
-            weights_to_download.extend(CLIPSEG_MODELS)
-
-    @staticmethod
-    def weights_map(base_url):
-        return {
-            model: {
-                "url": f"{base_url}/clipseg/{model}.tar",
-                "dest": "ComfyUI/models/clipseg",
-            }
-            for model in CLIPSEG_MODELS
-        }
+            weights_to_download.extend(["models--CIDAS--clipseg-rd64-refined"])
 
     @staticmethod
     def check_for_unsupported_nodes(node):