+++ scripts/clone_plugins.sh
@@ -32,6 +32,7 @@ repos=(
   "https://github.com/ZHO-ZHO-ZHO/ComfyUI-BRIA_AI-RMBG 44a3f8f"
   "https://github.com/Suzie1/ComfyUI_Comfyroll_CustomNodes d78b780"
   "https://github.com/huchenlei/ComfyUI-layerdiffuse 151f746"
+  "https://github.com/kijai/ComfyUI-KJNodes 48d5a18"
 )
 
 # Destination directory